move main view components to a views subfolder of views
