+++ resources/js/components/views/grid-bookmarks.tsx
